backend/ml_inference_fastapi_app/requirements.txt
fastapi
uvicorn[standard]

# ML Model Dependencies
torch
openai-clip
transformers
Pillow
numpy
python-multipart 